+++ src/lib/hooks/useVirtualizationTableHeight.tsx
@@ -10,7 +10,7 @@ const adjustTableHeight = (
 
   // calculate the height for the pseudo element after the table
   const existingPseudoElement = window.getComputedStyle(tableRef.current, "::after");
-  const existingPseudoHeight = parseFloat(existingPseudoElement.height) || 0;
+  const existingPseudoHeight = Number.parseFloat(existingPseudoElement.height) || 0;
   const tableHeight = tableRef.current.clientHeight - existingPseudoHeight;
   const pseudoHeight = Math.max(virtualHeight - tableHeight, 0);
   onPseudoHeightChange?.(pseudoHeight);