+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - added the `usePersistentReactDataTableState` hook which saves it's own state into the local storage
 
+### Changed
+
+- migrated to eslint 9 flat config
+- removed obsolete storybook
+
 ## [5.14.0] - 2025-12-15
 
 ### Fixed